allow radarr monitor to be all the values
In a definition, radarr_monitor can be movie/collection/none, but in the
config.yml where radarr is defined, monitor has to be a boolean. Sonarr
can take the full values in both places.

This just looks like a little oversight that can be fixed. (but I'm not
sure how it will affect existing true/false values)